Create Car_Records_DB_Init.py
This is a script that initializes a Database in SQLite that holds information on certain information about different vehicles. It uses class module call "car_class" that is unavailable to me at the time, so it does not work properly, but it gives a demonstration on how to initialize a database with Python using SQLite.